Encode / escape values that are added to XML doc
If a value is set on a Request that contains a character that has a
special meaning in XML, DOMDocument::createElementNS() triggers an error
if that value is passed in unhandled.

To handle this, instead of passing the value to createElementNS(),
it is let a bit later using the textContent property. This property
takes care of the encoding and escaping for us.

Fixes #5 Descriptions with a & cause a crash